ast/index_test: add cases covering 'every'
Expressions in the body of `every x, y in DOMAIN { BODY }` need to be
ignored: the overall statement could still be true even if BODY contains
undefined references: if the DOMAIN is empty.

The DOMAIN of every, however, will be subject to indexing. This happens
automatically because of previous rewriting steps, but the tests cases
still seemed like a good idea to include.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>