ast/parser: hint about 'every' future keyword
This should inject a hint for when it's likely that a parsed
identifier `every` was meant to be a keyword.

By the usual hinting mechanism in the parser, if the parser proceeds
without any errors, the hint is going to be ignored.

This doesn't look exactly like what we did for `some x in xs` because
the `some` keyword wasn't new (`every` is).

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>