ast/compile: don't error on unused declared+generated vars (every)
When compiling modules containing `every v in vs { ... }`, it will
get rewritten to `every __local0__, v in vs { ... }`.

Sending that module through the compiler _again_, it will complain
about the generated var being declared (via every, it declares its
key and val variables) but unused.

Now, we'll check if the original variable was already a generated
variable, and suppress the error in that case.

Fixes #4420.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>