download: Update regular polling fallback mechanism
OPA should fallback to regular polling either when the
server does not support long polling or the long polling
timeout is not specified by the user. This change adds
the latter. Without this change if OPA is started w/o
any polling config and if the server supports long polling,
OPA would incorrectly use long polling instead of the
execpted regular polling mode.

Signed-off-by: Ashutosh Narkar <anarkar4387@gmail.com>