topdown/http: parse (and cache) yaml response
This is similar to what we currently do for application/json responses, or when
force_json_decode is set to true: We parse the result, and according to how it's
configured, cache either the serialised or deserialised content.

Now, we'll also take care of yaml-related content-types, application/yaml and
application/x-yaml; and enable forcing yaml-decoding via force_yaml_decode.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>