topdown/tokens_test: adjust expected signature for go1.18.1
Go 1.18.1 was released last night -- and while it doesn't change
the signatures, it changes what `runtime.Version()` returns.

Our test expectation didn't account for that.

We've been pulling in go1.18.1 because our .go-version has 1.18 in it,
and the docker image ref built from that for the `ci-go-*` make targets
will use golang:1.18, which refers to the latest patch release.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>