Expose Capabilities through cli command
There is a new command argument i.e capabilities. With this, it is
possible to print the current capabilities version, show all
capabilities versions & print any capabilities version, without the need
of a file. Moreover, for the other commands which use the --capabilities
flag, it is possible to give only the version number, without specifying
a file. However, there are no breaking changes for the ones, who use the
capabilities file as an input for the flag. Unit tests were also
written, in order to test the new argument and the changes made in ast

Fixes: #4236
Signed-off-by: IoannisMatzaris <matzarisioannis@gmail.com>