ast/compile: move mocking arity check into type checker
This way, we're comparing function signatures of built-ins and non-built-ins
where we've already gathered them: during type checking.

Note that this doesn't introduce first-class functions as a completely new
concept, but technically, we're passing around functions as first-class
entities in the `with` replacements. So we better check them as such, too.

The changes to the ast/check.go code are covered by the existing cases, and
the added ones, in ast/compile_test.go.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>