topdown/net: require prefix length for IPv6 in net.cidr_merge
There are no default prefixes in IPv6, so if an IPv6 is fed into
net.cidr_merge, we'll return a non-halt error now.

Before, we'd fail in various ways if a prefix-less IPv6 was fed into
`net.cidr_merge`. With only one, we'd return `[ "<nil>" ]`, with two,
we'd panic.

Fixes #4596.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>