Allow disabling util.Roundtrip on Write
Add option to inmem.store which allows disabling the round-tripping
through JSON when adding data to the store.

This option is intended for callers who can guarantee the objects they
pass to Write are JSON objects, and have properly ensured the object
will be only be accessed by store once added.

See #4708

Signed-off-by: Will Beason <willbeason@google.com>