ci: pass --region to 'aws s3 sync' command
This is an attempt to fix an obscure condition that started last week:
the `aws s3 sync` command in our `deploy-ci` make target started failing,
with only this output:

    <botocore.awsrequest.AWSRequest object at 0x7f3198832f70>

A promising google hit suggests that it's got to do with "what's the best
region for me?" auto-discovery failing. Passing any region should then do
the trick.

Ref: https://florian.ec/blog/github-actions-awscli-errors/

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>