bundle: split large values in bundle iterator
Before, we'd failed to activate a bundle with a large data.json which
would end up being attempted to load in a single write.

Now, we'll recognize if the value is larger than a threshold, and split
it into multiple blobs, which hopefully go in with a single txn.

Activating the bundle in question works, but re-activating it afterwards
still fails in a different place: eraseBundles accumulates too many writes
in a single txn, attempting to reset the data in store.)

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>