bundle/store: deal with erasing large bundles
Before, we've accumulated too-large transactions by retrieving data keys
according to their partitioning, and deleting them one-by-one.

Now, we'll do the same thing, but in an iterator passed to Truncate, which
deals with restarting transactions.

This also means that we've got an extra call to Truncate, so one Truncate
operation more than before. However, we've been having one Truncate per
bundle already. To fix this, we could combine multiple storage.Iterator into
a chaining iterator, and only do one Truncate operation, but that's a future
optimization.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>