ci: remove quick-fuzz, guard wasm tests on changed files
This is a first step into running less things all the time that don't need to
be run all the time.

It's a heuristic, and as such fallible: there could always be changes that
break something in the wasm code path, because I have forgotten that there's a
dependency of some sort.

Removing the quick-fuzz target, it's never brought any issues up; and still
runs in nightly tests.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>