netlify: add debug output from 'ignore' script, simplify diff approach
`main` on netlify is a really old branch, and we have no way to update it.

So instead, we'll use what netlify intended: cached vs latest ref. Let's see
if we run into problems.

Concretely, it'll mean that if a PR updates parts of the website in one commit,
netlify will build the PR branch. If a subsequent commit to that PR branch does
not change anything in the website as it was in the previously built commit, no
build will happen. This could be confusing, it also makes sense: it was already
built, why build it again?

Also adds the build/ folder to the list of watched files: much of the docs
generation logic lives there.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>