ast/compile: move ref rewriting into earlier stage
Before, we had checked the refs early, and accepted a ref like

    p[y.z] { y := input }

and had that rule explode in the RuleTree, because it tried to use `y.z` as
key in a `map[ast.Value]*TreeNode`.

Now, we've moved the rewriting logic that rewrites the example rule to

    p[__local0__] { y := input; __local0__ := y.z }

to happen earlier. The rule tree should now only fed rule heads that have
hashable ref parts.

This required moving the InitLocalVarGen stage up the chain, but as it's still
below ResolveRefs, we should be OK.

In our internal backtesting results, this change should cover these panics:

    runtime error: hash of unhashable type ast.Ref

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>