oracle: cope with the rewriting we do for ref heads
The added test case passes as-is on main, the changes we do here are to
keep it that way.

It's been made necessary because of this dilemma:

1. The compiler rewrites rule head refs early because the rule tree expects
   only simple vars, no refs, in rule head refs. So `p[x.y]` becomes
   `p[local] { local = x.y }`
2. The oracle circles in on the node it's finding the definition for based
   on source location, and the logic for doing that depends on unaltered
   modules.

With this, (2.) is relaxed: the logic for building the lookup node stack can
now cope with generated statements that have been appended to the rule bodies.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>