bundle: Remove raw bytes check for lazy bundle loading mode
The raw bytes on the bundle object are only set if the bundle
contains any policy or data. So while activating a bundle w/o
policy or data in lazy mode we check if the raw bytes are set
and generate an error if not. This check would prevent valid
bundles from being activated if they contained no policy/data.

Signed-off-by: Ashutosh Narkar <anarkar4387@gmail.com>