SECURITY: migrate policy to web site, update content
It's easier to link to, and easier to find that way. Also, it's deliberately
not part of the versioned docs: we don't want anyone to find old information
when it's important.

Fixes #4272.

Note that we won't win any style awards for this, but let's make CSS tweaking
a follow-up issue.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>