rego/rego: Fix unneeded NDBCache initializations.
The NDBCache is intended as an opt-in feature, and the initializations
present in the `rego` module meant that it was forced on virtually every
eval. This commit removes those initializations, so that the only way an
NDBCache will ever make it to the evaluator is if it's provided from
outside the evaluator.

Signed-off-by: Philip Conrad <philipaconrad@gmail.com>