rego: skip rules for parsing query
This is underlying the code flow for

    cat query | opa eval --stdin

and now does what you'd expect: treat the input as query, and parse it as such.

Previously, we'd been using the parser opts from the future imports helper, and
would have missed setting SkipRules.

This seems to be the only non-test caller of ParseBodyWithOpts.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>