topdown/eval: skip default functions in full extent (#5259)
The previous code in evalVirtualComplete had the intention of skipping
functions, but failed to cature the `default f(x) := 1` case.

It's somewhat peculiar, since the syntax is accepted, but there is no eval
logic supporting it in topdown. (IR/Wasm support default functions.)

Either way, a function definition alone should not anything to the extent of
a package.

Fixes #5202.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>
Signed-off-by: Byron Lagrone <byron.lagrone@seqster.com>