topdown/http: remove socket param for unix sockets (#5355)
Removes the `socket` parameter from any http request targeting a UNIX domain
socket. The socket path is handled by OPA to identify which socket to
communicate with. It serves no purpose to send it to the socket itself. In some
cases, the parameter may also be rejected by the listening program.

All the `socket` parameter values are removed to prevent HTTP parameter
pollution.

The implementation may change the overall query parameters order. But there is
no standard regarding that so it should be okay.

Fixes #5313

Signed-off-by: Tuan Le <webmaster@michivi.com>