topdown/copyprop: skip 'every' body when re-adding removed eqs
Follow-up to 232b0c76e4de6202d2920e0059957461aaccbe84.

For the following policy, the second expression does make a difference:

    every x in input.xs { x = input.foo }
    _ = input.foo

Although it's part of the `every` body, when `input.xs` is `[]`, the body is
not evaluated and the overall `every` expression is true. So, we cannot drop
that second equation because its ref already is in the body of `every`.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>