ast/lazyObj.Insert(): force on insert, don't panic
So with the panic() we've had there before, we've established that in our code,
the Insert() won't happen, so lazyObj can be treated as immutable.

However, it's conceivable that someone is using OPA-via-Golang, and calls
Insert on a lazyObj returned from a topdown query. While there are opt-out
options, it's still nicer satisfy the entire ast.Object interface if we can.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>