docs: Document reference timestamp behavior for `time.parse_ns`
This commit adds documentation and examples around one of the bigger
"gotchas" in handling time values in OPA: parsing date formats.

The issue is that users of `time.parse_ns` have to use the same
"reference timestamp" values as Go's `time/format` module does, which
can be deeply surprising behavior to newcomers to the Go language,
and to OPA users as well.

By documenting this, the hope is that we'll help users make fewer
mistakes around parsing date formats in OPA.

Signed-off-by: Philip Conrad <philipaconrad@gmail.com>