providers/aws: add AWS V4 Asymmetric (V4A) signing
AWS is rolling out an extension to SigV4 called Signature Version 4A (SigV4A)
which enables signatures that are valid in more than one AWS Region. This is
required for signing multi-region API requests, for example with Amazon S3
Multi-Region Access Points (MRAP). This lets OPA use an S3 MRAP as a bundle
source that is latency optimized by geography and also highly-available /
resilient against regional outages.

This commit adds SigV4A support with an implementation that is a modified
version of internal code from the aws-sdk-go-v2 project:
https://github.com/aws/aws-sdk-go-v2/tree/93c3f18/internal/v4a. That approach
felt far safer than trying to re-implement the necessary crypto functions
from scratch. Note: the aws-sdk-go-v2 project is Apache 2 licensed, so
copying and modifying their code seems fine according to section 4. I've
attempted to satisfy the conditions of that section by:
* 4.b: adding a comment line at the top of
  internal/providers/aws/signing_v4a.go
* 4.d: including the NOTICES.txt file from aws-sdk-go-v2 in the
  internal/providers/aws/ folder

This also makes an internal/providers/aws package that contains both the
existing V4 signing code as well as the V4A signing code being added.
This seemed reasonable for code organization and because this is
internal code.

Fixes #5429

Signed-off-by: Jay Wineinger <jawineinger@spscommerce.com>